Description for User: - responses type updates; - DTOs updates; - added description in Users API methods.
